cleaned up how fakes are set up.
