+++ package.json
@@ -93,6 +93,7 @@
     "react-test-renderer": "^17.0.2",
     "ts-node": "^9.1.1",
     "tsconfig-paths-webpack-plugin": "^3.5.1",
+    "type-fest": "^2.1.0",
     "typescript": "^4.2.4"
   },
   "resolutions": {
