+++ src/test-utils/fakes/ua/listColumns.ts
@@ -1,143 +1,6 @@
-// Copyright 2020 Google Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+import { Column } from "@/types/ua"
 
-import * as React from "react"
-import { Provider } from "react-redux"
-import { makeStore } from "../gatsby/wrapRootElement"
-import {
-  createHistory,
-  createMemorySource,
-  LocationProvider,
-  History,
-} from "@reach/router"
-import { QueryParamProvider } from "use-query-params"
-import { AccountSummary, Column } from "./types/ua"
-
-interface WithProvidersConfig {
-  path?: string
-  isLoggedIn?: boolean
-  uaListAccountSummaries?: () => Promise<gapi.client.analytics.AccountSummary[]>
-}
-
-export const wrapperFor: (options: {
-  path?: string
-  isLoggedIn?: boolean
-  setUp?: () => void
-}) => React.FC = ({ path, isLoggedIn, setUp }) => {
-  path = path || "/"
-  isLoggedIn = isLoggedIn === undefined ? true : isLoggedIn
-
-  window.localStorage.clear()
-  setUp && setUp()
-
-  const history = createHistory(createMemorySource(path))
-  const store = makeStore()
-
-  if (isLoggedIn) {
-    store.dispatch({ type: "setUser", user: {} })
-  } else {
-    store.dispatch({ type: "setUser", user: undefined })
-  }
-
-  const Wrapper: React.FC = ({ children }) => (
-    <Provider store={store}>
-      <LocationProvider history={history}>
-        <QueryParamProvider {...{ default: true }} reachHistory={history}>
-          {children}
-        </QueryParamProvider>
-      </LocationProvider>
-    </Provider>
-  )
-  return Wrapper
-}
-export const TestWrapper = wrapperFor({})
-
-export const withProviders = (
-  component: JSX.Element | null,
-  { path, isLoggedIn, uaListAccountSummaries }: WithProvidersConfig = {
-    path: "/",
-    isLoggedIn: true,
-  }
-): {
-  wrapped: JSX.Element
-  history: History
-  store: any
-  gapi: ReturnType<typeof testGapi>
-} => {
-  path = path || "/"
-  isLoggedIn = isLoggedIn === undefined ? true : isLoggedIn
-
-  window.localStorage.clear()
-
-  const history = createHistory(createMemorySource(path))
-  const store = makeStore()
-
-  if (isLoggedIn) {
-    store.dispatch({ type: "setUser", user: {} })
-  } else {
-    store.dispatch({ type: "setUser", user: undefined })
-  }
-
-  const gapi = testGapi({ uaListAccountSummaries })
-  store.dispatch({ type: "setGapi", gapi })
-
-  const wrapped = (
-    <Provider store={store}>
-      <LocationProvider history={history}>
-        <QueryParamProvider reachHistory={history}>
-          {component}
-        </QueryParamProvider>
-      </LocationProvider>
-    </Provider>
-  )
-  return { wrapped, history, store, gapi }
-}
-
-const testAccounts: AccountSummary[] = [
-  {
-    id: "account-id-1",
-    name: "Account Name 1",
-    webProperties: [
-      {
-        id: "property-id-1-1",
-        name: "Property Name 1 1",
-        profiles: [
-          { id: "view-id-1-1-1", name: "View Name 1 1 1" },
-          { id: "view-id-1-1-2", name: "View Name 1 1 2" },
-        ],
-      },
-      {
-        id: "property-id-1-2",
-        name: "Property Name 1 2",
-        profiles: [{ id: "view-id-1-2-1", name: "View Name 1 2 1" }],
-      },
-    ],
-  },
-  {
-    id: "account-id-2",
-    name: "Account Name 2",
-    webProperties: [
-      {
-        id: "property-id-2-1",
-        name: "Property Name 2 1",
-        profiles: [{ id: "view-id-2-1-1", name: "View Name 2 1 1" }],
-      },
-    ],
-  },
-]
-
-const testColumns: Column[] = [
+const columns: Column[] = [
   {
     id: "ga:userType",
     kind: "analytics#column",
@@ -419,111 +282,8 @@ const testColumns: Column[] = [
   },
 ]
 
-const listPromise = Promise.resolve({ result: { items: testAccounts } })
-// TODO - add in real type.
-const metadataColumnsPromise = Promise.resolve({
-  result: {
-    items: testColumns,
-    etag: "abc123",
-  },
-})
+const listColumns: typeof gapi.client.analytics.metadata.columns.list = () => {
+  return Promise.resolve({ result: { items: columns } }) as any
+}
 
-export const testGapi = ({
-  uaListAccountSummaries,
-}: WithProvidersConfig | undefined = {}) => ({
-  client: {
-    analyticsadmin: {
-      properties: {
-        iosAppDataStreams: {
-          list: (): Promise<{
-            result: gapi.client.analyticsadmin.GoogleAnalyticsAdminV1alphaListIosAppDataStreamsResponse
-          }> => Promise.resolve({ result: {} }),
-        },
-        androidAppDataStreams: {
-          list: (): Promise<{
-            result: gapi.client.analyticsadmin.GoogleAnalyticsAdminV1alphaListAndroidAppDataStreamsResponse
-          }> => Promise.resolve({ result: {} }),
-        },
-        webDataStreams: {
-          list: (): Promise<{
-            result: gapi.client.analyticsadmin.GoogleAnalyticsAdminV1alphaListWebDataStreamsResponse
-          }> => Promise.resolve({ result: {} }),
-        },
-      },
-      accountSummaries: {
-        list: ({
-          pageToken,
-        }: {
-          pageToken?: string
-        }): Promise<{
-          result: gapi.client.analyticsadmin.GoogleAnalyticsAdminV1alphaListAccountSummariesResponse
-        }> => {
-          if (pageToken === "1") {
-            return Promise.resolve({
-              result: {
-                accountSummaries: [
-                  {
-                    account: "accounts/def456",
-                    displayName: "my second account",
-                    name: "accountSummaries/def",
-                    propertySummaries: [
-                      {
-                        displayName: "my third property",
-                        property: "properties/3",
-                      },
-                      {
-                        displayName: "my fourth property",
-                        property: "properties/4",
-                      },
-                    ],
-                  },
-                ],
-              },
-            })
-          }
-          return Promise.resolve({
-            result: {
-              accountSummaries: [
-                {
-                  account: "accounts/abc123",
-                  displayName: "my first account",
-                  name: "accountSummaries/abc",
-                  propertySummaries: [
-                    {
-                      displayName: "my first property",
-                      property: "properties/1",
-                    },
-                    {
-                      displayName: "my second property",
-                      property: "properties/2",
-                    },
-                  ],
-                },
-              ],
-              nextPageToken: "1",
-            },
-          })
-        },
-      },
-    },
-    analytics: {
-      management: {
-        accountSummaries: {
-          list: uaListAccountSummaries
-            ? () =>
-                uaListAccountSummaries().then(a => ({
-                  result: { items: a },
-                }))
-            : () => listPromise,
-        },
-      },
-      metadata: {
-        columns: {
-          list: () => {
-            return metadataColumnsPromise
-          },
-        },
-      },
-    },
-  },
-})
+export default listColumns
